# Conflicts:
#	Common/ac/gamestructdefines.h
#	Common/game/main_game_file.cpp
#	Common/gfx/allegrobitmap.cpp
#	Editor/AGS.Editor/AGSEditor.cs
#	Editor/AGS.Editor/AGSEditorController.cs
#	Editor/AGS.Editor/NativeProxy.cs
#	Editor/AGS.Editor/Panes/SpriteImportWindow.Designer.cs
#	Editor/AGS.Editor/Panes/SpriteImportWindow.cs
#	Editor/AGS.Editor/Panes/SpriteSelector.cs
#	Editor/AGS.Editor/Utils/SpriteTools.cs
#	Editor/AGS.Native/NativeMethods.cpp
#	Editor/AGS.Native/NativeMethods.h
#	Editor/AGS.Native/agsnative.cpp
#	Editor/AGS.Types/Sprite.cs
#	Engine/ac/character.h
#	Engine/gfx/ali3dogl.cpp
#	Engine/gfx/ali3dogl.h

+++ Editor/AGS.Editor/AGSEditor.cs
@@ -114,7 +114,8 @@ namespace AGS.Editor
          *                  GlobalVariable may be of array type.
          * 3.6.2.2        - Button.WrapText, TextPadding.
          * 3.6.2.6        - Settings.GameFPS.
-         * 3.6.3          - Settings.GUIHandleOnlyLeftMouseButton
+         * 3.6.2.9        - Sprite.TransparentColorIndex (can select transparent palette index).
+         * 3.6.3          - Settings.GUIHandleOnlyLeftMouseButton.
          * 
          * 3.99.99.00     - BlendMode for various objects, Character.Transparency.
          * 3.99.99.01     - Open rooms
@@ -575,6 +576,9 @@ namespace AGS.Editor
 
         public void RenameFileOnDisk(string currentName, string newName)
         {
+            if (!File.Exists(currentName))
+                return; // nothing to rename
+
             string sourcePath = Path.GetFullPath(currentName);
             string destPath = Path.GetFullPath(newName);
             File.Move(sourcePath, destPath);
